requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent==23.9.1

# ML/NLP
sentence-transformers==2.2.2
transformers==4.41.2
torch==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
tiktoken==0.7.0
nltk==3.8.1

# PDF processing
pypdf==4.2.0
# OR pymupdf==1.23.26  # Alternative, better for complex PDFs

# Text splitting
langchain-text-splitters==0.2.2

# Vector DB
pinecone-client==3.0.0

# LLM
anthropic==0.34.2

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
pandas==2.2.2